--- src/pages/Public/Login.tsx
@@ -1,27 +1,0 @@
-import styled from 'styled-components';
-import LoginAndRegister from '@/components/Organisms/LoginAndRegister';
-import Logo from '@/components/Atoms/Logo';
-
-const StyledLogin = styled.div`
-  ${({ theme }) => theme.MIXIN.FLEX({ direction: 'column', align: 'center', justify: 'center' })};
-  width: 100%;
-  height: 100vh;
-
-  & > a:first-child {
-    margin-bottom: 64px;
-
-    svg {
-      width: 100%;
-      height: 100%;
-    }
-  }
-`;
-
-const Login = () => (
-  <StyledLogin>
-    <Logo logoSize="Large" />
-    <LoginAndRegister />
-  </StyledLogin>
-);
-
-export default Login;
